refactor: fixed airline view file name
